vitamin-p/requirements.txt
tensorflow==2.6.0
numpy==1.19.5
scikit-learn==0.24.2
opencv-python==4.5.3.56
pyyaml==5.4.1
matplotlib==3.4.3
pandas==1.3.3
